+++ scripts/tests/harness.sh
@@ -5,6 +5,7 @@
 
 FOREST_PATH="forest"
 FOREST_CLI_PATH="forest-cli"
+FOREST_WALLET_PATH="forest-wallet"
 FOREST_TOOL_PATH="forest-tool"
 
 TMP_DIR=$(mktemp --directory)
@@ -12,6 +13,7 @@ LOG_DIRECTORY=$TMP_DIR/logs
 
 export TMP_DIR
 export LOG_DIRECTORY
+export FOREST_WALLET_PATH
 
 function forest_download_and_import_snapshot {
   echo "Downloading and importing snapshot"