+++ src/cli/subcommands/mod.rs
@@ -21,7 +21,6 @@ mod shutdown_cmd;
 mod snapshot_cmd;
 mod state_cmd;
 mod sync_cmd;
-mod wallet_cmd;
 
 use std::io::{self, Write};
 
@@ -41,7 +40,7 @@ pub(super) use self::{
     car_cmd::CarCommands, chain_cmd::ChainCommands, config_cmd::ConfigCommands, db_cmd::DBCommands,
     mpool_cmd::MpoolCommands, net_cmd::NetCommands, send_cmd::SendCommand,
     shutdown_cmd::ShutdownCommand, snapshot_cmd::SnapshotCommands, state_cmd::StateCommands,
-    sync_cmd::SyncCommands, wallet_cmd::WalletCommands,
+    sync_cmd::SyncCommands,
 };
 use crate::cli::subcommands::info_cmd::InfoCommand;
 
@@ -68,6 +67,51 @@ pub struct FetchCommands {
     params_size: Option<String>,
 }
 
+// Those subcommands are hidden and only here to help users migrating to forest-wallet
+#[derive(Debug, clap::Subcommand)]
+pub enum WalletCommands {
+    New {
+        #[arg(default_value = "secp256k1")]
+        signature_type: String,
+    },
+    Balance {
+        address: String,
+    },
+    Default,
+    Export {
+        address: String,
+    },
+    Has {
+        key: String,
+    },
+    Import {
+        path: Option<String>,
+    },
+    List {
+        #[arg(long, alias = "exact-balance", short_alias = 'e')]
+        no_round: bool,
+        #[arg(long, alias = "fixed-unit", short_alias = 'f')]
+        no_abbrev: bool,
+    },
+    SetDefault {
+        key: String,
+    },
+    Sign {
+        #[arg(short)]
+        message: String,
+        #[arg(short)]
+        address: String,
+    },
+    Verify {
+        #[arg(short)]
+        address: String,
+        #[arg(short)]
+        message: String,
+        #[arg(short)]
+        signature: String,
+    },
+}
+
 /// Forest binary sub-commands available.
 #[derive(clap::Subcommand, Debug)]
 pub enum Subcommand {
@@ -87,7 +131,8 @@ pub enum Subcommand {
     #[command(subcommand)]
     Net(NetCommands),
 
-    /// Manage wallet
+    // Those subcommands are hidden and only here to help users migrating to forest-wallet
+    #[command(hide = true)]
     #[command(subcommand)]
     Wallet(WalletCommands),