+++ CHANGELOG.md
@@ -40,6 +40,8 @@
 - [#3355](https://github.com/ChainSafe/forest/pull/3355) Moved commands
   - `forest-cli db stats` to `forest-tool db stats`
   - `forest-cli db clean` to `forest-tool db destroy`
+- [#3362](https://github.com/ChainSafe/forest/pull/3362) Moved the following
+  `forest-cli wallet` subcommands to `forest-wallet`
 
 ### Added